fix: directory structure
